requirements.txt
flask==2.0.2
Werkzeug==2.0.3
jinja2==3.0.2
flask_login==0.5.0
flask_migrate==3.1.0
WTForms==3.0.0
flask_wtf==1.0.0
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.29
email_validator==1.1.3
gunicorn==20.1.0
flask-restx==0.5.1
python-dotenv==0.19.2
Flask-Minify==0.37
Flask-Dance==5.1.0
blinker==1.4
pyOpenSSL

flask_mysqldb
flask_mail
numpy==1.22.4
pandas==1.4.2; python_version >= "3.8"
plotly==5.8.0; python_version >= "3.6"
# psycopg2-binary
